Minor optimization to TextDocumentStates.ComputeFilePathToDocumentIds
This method commonly shows up in CPU investigations of roslyn scenarios. The Add case is *significantly* more common than the modify existing case, so this optimizes the add case to not do both the TryGetValue and calling the setter. (Each of these show up as about 20% of the CPU time spent in this method) Although minor,this would shave off about 80ms of CPU time (0.3%) in the profile I'm looking at.