*** WIP: Modify VSInternalContinueCharacter derived class serialization
Currently, the VSInternalContinueCharacter derived classes don't serialize the _vs_type property or have the Kind attribute specified on the classes. This needs to be done both on the Roslyn side and the vslanguageserverclient side.

This will greatly reduce the serialization exception count seen in a razor speedometer test due to the complexity of the SumType in VSInternalCompletionList referencing these types.